Abstract offset-based and pointer-based buffers.
The pointer-based buffers basically function exactly like regular
syscalls do; the user passes slices of data to the kernel, which then
handles the syscall directly, or inserts a grant and lets the server
handle it.

The offset-based on the other hand, works with buffer pools (mainly
paired with redox-buffer-pool to make it easier). Instead of passing an
address and the length to the data, it'll pass the offset within the
buffer pool (pretend the buffer pool is a file, and the offset is the
mmap offset in that file), and the length instead.